Reorganise Demex Products components and remove pools sagas
