+++ src/@demex-info/store/actions.ts
@@ -1,15 +1,11 @@
 import * as App from "./app/actions";
 import * as Layout from "./layout/actions";
 import * as Markets from "./markets/actions";
-import * as Pools from "./pools/actions";
-import * as Staking from "./staking/actions";
 
 const actions = {
 	App,
 	Layout,
   Markets,
-  Pools,
-  Staking,
 };
 
 export default actions;
