+++ src/@demex-info/utils/index.ts
@@ -3,5 +3,6 @@ export * from "./date";
 export * from "./liquidityPool";
 export { default as logger } from "./logger";
 export * from "./number";
+export * from "./staking";
 export * from "./strings";
 export * from "./tendermint";
