+++ src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/LiquidityPool/index.ts
@@ -1,0 +1,1 @@
+export { default as LiquidityPool } from "./LiquidityPool";
