--- src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/ProductCarousel/components/index.ts
@@ -1,2 +1,0 @@
-export { default as LiquidityPoolSlide } from "./LiquidityPoolSlide";
-export { default as StakingSlide } from "./StakingSlide";
