--- src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/ProductCarousel/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductCarousel } from "./ProductCarousel";
