--- src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/ProductScroll/components/index.ts
@@ -1,2 +1,0 @@
-export { default as LiquidityPoolSection } from "./LiquidityPoolSection";
-export { default as StakingSection } from "./StakingSection";
