--- src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/ProductScroll/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductScroll } from "./ProductScroll";
