+++ src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/Staking/StakingSection.tsx
@@ -2,12 +2,8 @@ import { Staking } from "@demex-info/assets";
 import { RenderGuard, TypographyLabel } from "@demex-info/components";
 import { getDemexLink, goToLink, lottieDefaultOptions, Paths } from "@demex-info/constants";
 import { useTaskSubscriber } from "@demex-info/hooks";
-import { StakingTasks } from "@demex-info/store/staking/types";
 import { RootState } from "@demex-info/store/types";
-import {
-  BN_ZERO, parseNumber, SECONDS_PER_HOUR, SECONDS_PER_MINUTE,
-  SECONDS_PER_YEAR, toPercentage, toShorterNum,
-} from "@demex-info/utils";
+import { BN_ZERO, toPercentage, toShorterNum, StakingStats } from "@demex-info/utils";
 import { Box, Button, Divider, makeStyles, Theme, Typography } from "@material-ui/core";
 import { Skeleton } from "@material-ui/lab";
 import BigNumber from "bignumber.js";
@@ -17,8 +13,17 @@ import React, { useEffect } from "react";
 import { useInView } from "react-intersection-observer";
 import { useSelector } from "react-redux";
 
-const StakingSection: React.FC = () => {
+interface DataProps {
+  apr: BigNumber;
+  stats: StakingStats;
+}
+
+const StakingSection: React.FC<DataProps> = (props: DataProps) => {
+  const { apr, stats } = props;
   const classes = useStyles();
+  const [loading] = useTaskSubscriber("runStaking");
+
+  const network = useSelector((state: RootState) => state.app.network);
 
   const lottieRef = React.useRef<any>();
   const [stakingTxtRef, stakingTxtView] = useInView({
@@ -29,21 +34,6 @@ const StakingSection: React.FC = () => {
     threshold: 0.2,
     triggerOnce: true,
   });
-
-  const [statsLoading] = useTaskSubscriber(StakingTasks.Stats);
-  const [aprLoading] = useTaskSubscriber(StakingTasks.Blocks, StakingTasks.AvgBlockTime, StakingTasks.Validators);
-
-  const network = useSelector((state: RootState) => state.app.network);
-  const { avgBlockTime, avgReward, stats, totalBonded } = useSelector((state: RootState) => state.staking);
-
-  const timeArray: any = avgBlockTime.split(":");
-  const hours: BigNumber = parseNumber(timeArray[0], BN_ZERO)!.times(SECONDS_PER_HOUR);
-  const minutes: BigNumber = parseNumber(timeArray[1], BN_ZERO)!.times(SECONDS_PER_MINUTE);
-  const seconds: BigNumber = parseNumber(timeArray[2], BN_ZERO)!;
-  const blockTimeBN: BigNumber = hours.plus(minutes).plus(seconds);
-  const blocksInYear = new BigNumber(SECONDS_PER_YEAR).div(blockTimeBN);
-  const rewardsInYear = blocksInYear.times(avgReward);
-  const apr = totalBonded.isZero() ? BN_ZERO : rewardsInYear.div(totalBonded);
 
   const delayAnimation = () => {
     lottieRef?.current?.pause();
@@ -81,12 +71,12 @@ const StakingSection: React.FC = () => {
             <TypographyLabel color="textSecondary">
               Total Staked
             </TypographyLabel>
-            <RenderGuard renderIf={statsLoading}>
+            <RenderGuard renderIf={loading}>
               <Box>
                 <Skeleton width="5rem" height="3rem" />
               </Box>
             </RenderGuard>
-            <RenderGuard renderIf={!statsLoading}>
+            <RenderGuard renderIf={!loading}>
               <Typography variant="h4" color="textPrimary">
                 {toShorterNum(stats.totalStaked ?? BN_ZERO)} SWTH
               </Typography>
@@ -96,12 +86,12 @@ const StakingSection: React.FC = () => {
             <TypographyLabel color="textSecondary">
               Staking APR
             </TypographyLabel>
-            <RenderGuard renderIf={aprLoading}>
+            <RenderGuard renderIf={loading}>
               <Box>
                 <Skeleton width="5rem" height="3rem" />
               </Box>
             </RenderGuard>
-            <RenderGuard renderIf={!aprLoading}>
+            <RenderGuard renderIf={!loading}>
               <Typography variant="h4" color="textPrimary">
                 {toPercentage(apr, 2)}%
               </Typography>
