+++ src/@demex-info/views/MiddleBlackSection/components/DemexProducts/components/Staking/index.ts
@@ -1,0 +1,1 @@
+export { default as Staking } from "./Staking";
